refactor: Improve API code
